refactor(navbar): separate select into its own component
