+++ client/src/App.jsx
@@ -1,5 +1,5 @@
 import { SheetContextProvider } from "./context/SheetContext";
-import Navbar from "./components/Navbar";
+import Navbar from "./components/Navbar/Navbar";
 import Hero from "./components/Hero";
 import Spotify from "./components/Spotify/Spotify";
 import TestimonialSlider from "./components/Testimonials/TestimonialSlider";
